+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # CHANGELOG
 
+* `2.2.10` Release - [2.2.10](#2210)
 * `2.2.00` Release - [2.2.00](#2200)
 * `0.93.0` Release - [0.93.0](#0930)
 * `0.80.0` Release - [0.80.0](#0800)
 * `0.79.0` Release - [0.79.0](#0790)
 * `0.78.0` Release - [0.78.0](#0780)
 * `0.77.x` Releases - [0.77.0](#0770)
+
+## 2.2.100
+#### Features
+* `EMP-14806` Update support for downloading additional media : audio & subtitles
+
+#### Changes
+* `EMP-14806`  Updated to Swift 5
+* `EMP-14806`  Now the Exposure download module support iOS 11 & up versions 
 
 ## 2.2.000
 #### Features